moved buttons into compact leaderboard component and some details
