+++ src/app/(white)/admin/components/CompactLeaderboard.tsx
@@ -1,6 +1,4 @@
 import React, { useEffect, useState } from 'react'
-import { PlayerScore } from '@/lib/types/types'
-import { getActiveScores } from '@/lib/api/scoreApi'
 import Leaderboard from './Leaderboard'
 import { useEventName } from '@/lib/hooks/useEventName'
 import { BreadcrumbItem } from '@entur/menu'
@@ -8,27 +6,60 @@ import Link from 'next/link'
 import { useStopPlaceNames } from '@/lib/hooks/useStopPlaceName'
 import { TravelHeader } from '@entur/travel'
 import { BannerAlertBox } from '@entur/alert'
+import useScores from '@/lib/hooks/useScores'
+import { endActiveEvent, getActiveEvent } from '@/lib/api/eventApi'
+import { Modal } from '@entur/modal'
+import { Paragraph } from '@entur/typography'
+import { Button, SecondaryButton } from '@entur/button'
+import { useRouter } from 'next/navigation'
 
 const CompactLeaderboardPage: React.FC = (): JSX.Element => {
-    const [scores, setScores] = useState<PlayerScore[]>([])
-    const { isEventNameError } = useEventName()
+    const { isEventNameError, setEventNameError } = useEventName()
     const { startLocationName, endLocationName } = useStopPlaceNames()
+    const { scores, leader, setShowAlert } = useScores()
+    const router = useRouter()
+    const [isOpen, setOpen] = useState(false)
+    const [isWinnerEndOpen, setWinnerEndOpen] = useState(false)
+    const [isWinnerOpen, setWinnerOpen] = useState(false)
+    const [isActiveEvent, setIsActiveEvent] = useState<boolean>(false)
+
+    const checkActiveEvent = async (): Promise<boolean> => {
+        const result = await getActiveEvent()
+        return result?.isActive ?? false
+    }
 
     useEffect(() => {
-        const fetchScores = async () => {
-            const scores = await getActiveScores()
-            if (scores && scores.length > 0) {
-                const sortedScores = scores.sort(
-                    (a, b) =>
-                        b.scoreValue - a.scoreValue ||
-                        a.totalTravelTime - b.totalTravelTime,
-                )
-                setScores(sortedScores)
-            }
+        const fetchActiveEvent = async () => {
+            const isActive = await checkActiveEvent()
+            setIsActiveEvent(isActive)
         }
-
-        fetchScores()
+        fetchActiveEvent()
     }, [])
+
+    const handleDrawWinnerAndEndGame = async () => {
+        if (scores.length === 0) {
+            setShowAlert(true)
+        } else {
+            setWinnerEndOpen(false)
+            setWinnerOpen(true)
+            await endActiveEvent()
+        }
+    }
+
+    const handleEndGame = async () => {
+        const result = await endActiveEvent()
+        if (result.success) {
+            setOpen(false)
+            setEventNameError(true)
+            router.refresh
+        }
+    }
+
+    const handleDismiss = () => {
+        setWinnerOpen(false)
+        setEventNameError(true)
+        router.refresh
+    }
 
     return (
         <div>
@@ -41,8 +72,8 @@ const CompactLeaderboardPage: React.FC = (): JSX.Element => {
                     Gå til “Opprett spill” for å opprette et nytt spill
                 </BannerAlertBox>
             ) : (
-                <div className="flex flex-col pb-4">
-                    <div className="bg-white rounded shadow-md p-6">
+                <div className="flex pb-4 gap-16">
+                    <div className="bg-white rounded shadow-md p-6 min-w-[848px]">
                         <TravelHeader
                             size="large"
                             from={startLocationName}
@@ -66,6 +97,96 @@ const CompactLeaderboardPage: React.FC = (): JSX.Element => {
                             </BreadcrumbItem>
                         </div>
                     </div>
+                    <div className="flex flex-col max-w-md justify-items-end">
+                        <Modal
+                            open={isWinnerEndOpen}
+                            onDismiss={() => setWinnerEndOpen(false)}
+                            title="Trekk vinner og avslutt spill?"
+                            size="medium"
+                        >
+                            <Paragraph>
+                                Når du trekker en vinner avsluttes spillet
+                                automatisk. Det vil være mulig å gjenåpne
+                                spillet igjen på et senere tidspunkt.
+                            </Paragraph>
+                            <div className="flex gap-4">
+                                <Button
+                                    variant={'primary'}
+                                    className="max-w-[250px]"
+                                    onClick={handleDrawWinnerAndEndGame}
+                                    type="button"
+                                >
+                                    Trekk vinner og avslutt
+                                </Button>
+                                <SecondaryButton
+                                    className="w-[81px]"
+                                    onClick={() => setWinnerEndOpen(false)}
+                                >
+                                    Avbryt
+                                </SecondaryButton>
+                            </div>
+                        </Modal>
+                        <Button
+                            variant={'success'}
+                            size="large"
+                            className="max-w-[250px]"
+                            onClick={() => setWinnerEndOpen(true)}
+                            type="button"
+                        >
+                            Trekk vinner og avslutt spill
+                        </Button>
+                        <Paragraph margin="none" className="mt-2">
+                            Vinneren trekkes tilfeldig blant spillerne med
+                            høyest poengsum.
+                        </Paragraph>
+                        <Modal
+                            open={isOpen}
+                            onDismiss={() => setOpen(false)}
+                            title="Avslutt spill uten å trekke vinner?"
+                            size="medium"
+                        >
+                            <Paragraph>
+                                Du er i ferd med å avslutte spillet uten å trekk
+                                een vinner. Det vil være mulig å gjenåpne
+                                spillet og trekke en vinner på et senere
+                                tidspunkt.
+                            </Paragraph>
+                            <div className="flex items-center gap-4">
+                                <Button
+                                    variant="primary"
+                                    className="max-w-[250px]"
+                                    disabled={!isActiveEvent}
+                                    onClick={handleEndGame}
+                                >
+                                    Avslutt spill
+                                </Button>
+                                <Button
+                                    variant="secondary"
+                                    onClick={() => setOpen(false)}
+                                >
+                                    Avbryt
+                                </Button>
+                            </div>
+                        </Modal>
+                        <Button
+                            variant="secondary"
+                            size="large"
+                            className="max-w-[250px] mt-6"
+                            onClick={() => setOpen(true)}
+                            type="button"
+                        >
+                            Avslutt spill
+                        </Button>
+                    </div>
+                    <Modal
+                        open={isWinnerOpen}
+                        onDismiss={handleDismiss}
+                        title={`Vinner: ${leader?.player.playerName}`}
+                        size="medium"
+                    >
+                        <p>E-post: {leader?.player.email}</p>
+                        <p>Telefon: {leader?.player.phoneNumber}</p>
+                    </Modal>
                 </div>
             )}
         </div>
